Moved deprecated APIs to the obsolete PortalAliasController
